fix(driver): move redundant RawOp code
