+++ compio-driver/src/unix/mod.rs
@@ -9,73 +9,13 @@ use compio_buf::BufResult;
 
 use crate::{OpCode, PushEntry};
 
-pub(crate) struct RawOp {
-    op: NonNull<dyn OpCode>,
-    // The two flags here are manual reference counting. The driver holds the strong ref until it
-    // completes; the runtime holds the strong ref until the future is dropped.
-    cancelled: bool,
-    result: PushEntry<Option<Waker>, io::Result<usize>>,
+#[repr(transparent)]
+pub(crate) struct Overlapped<T: ?Sized> {
+    pub op: T,
 }
 
-impl RawOp {
-    pub(crate) fn new(_user_data: usize, op: impl OpCode + 'static) -> Self {
-        let op = Box::new(op);
-        Self {
-            op: unsafe { NonNull::new_unchecked(Box::into_raw(op as Box<dyn OpCode>)) },
-            cancelled: false,
-            result: PushEntry::Pending(None),
-        }
-    }
-
-    pub fn as_pin(&mut self) -> Pin<&mut dyn OpCode> {
-        unsafe { Pin::new_unchecked(self.op.as_mut()) }
-    }
-
-    pub fn set_cancelled(&mut self) -> bool {
-        self.cancelled = true;
-        self.has_result()
-    }
-
-    pub fn set_result(&mut self, res: io::Result<usize>) -> bool {
-        if let PushEntry::Pending(Some(w)) =
-            std::mem::replace(&mut self.result, PushEntry::Ready(res))
-        {
-            w.wake();
-        }
-        self.cancelled
-    }
-
-    pub fn has_result(&self) -> bool {
-        self.result.is_ready()
-    }
-
-    pub fn set_waker(&mut self, waker: Waker) {
-        if let PushEntry::Pending(w) = &mut self.result {
-            *w = Some(waker)
-        }
-    }
-
-    /// # Safety
-    /// The caller should ensure the correct type.
-    ///
-    /// # Panics
-    /// This function will panic if the result has not been set.
-    pub unsafe fn into_inner<T: OpCode>(self) -> BufResult<usize, T> {
-        let mut this = ManuallyDrop::new(self);
-        let op = *Box::from_raw(this.op.cast().as_ptr());
-        BufResult(
-            std::mem::replace(&mut this.result, PushEntry::Pending(None))
-                .take_ready()
-                .unwrap(),
-            op,
-        )
-    }
-}
-
-impl Drop for RawOp {
-    fn drop(&mut self) {
-        if self.has_result() {
-            let _ = unsafe { Box::from_raw(self.op.as_ptr()) };
-        }
+impl<T> Overlapped<T> {
+    pub(crate) fn new(_driver: RawFd, _user_data: usize, op: T) -> Self {
+        Self { op }
     }
 }
